[TimeInterval] Moved it to BaseLib
