+++ ProcessLib/BoundaryCondition/CreateBoundaryCondition.cpp
@@ -22,7 +22,7 @@
 #include "PhaseFieldIrreversibleDamageOracleBoundaryCondition.h"
 #include "RobinBoundaryCondition.h"
 
-#include "NumLib/TimeStepping/TimeInterval.h"
+#include "BaseLib/TimeInterval.h"
 
 #ifdef OGS_USE_PYTHON
 #include "Python/PythonBoundaryCondition.h"
